Reduce database queries executed when using `our-img` tag helper
`our-img` tag helper is executing a database query each time it's used
due to using `IMediaService` to fetch the media item. Upon inspection
this call is redundant as the properties (`umbracoHeight` and
`umbracoWidth`) can be accessed via the `MediaItem` that is passed in as
a parameter.

This change did result in a divide by zero exception occuring, which was
a bit of a head scratcher as fetching the height & width via the media
service also resulted in dividing by zero however no exception was
thrown. Added some code to prevent a divide by zero occuring when the
image height isn't known (which is typically the case for SVG files).